gui: split components into separate files
